requirements.txt
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
coveralls==1.2.0
defusedxml==0.5.0
Django==2.0.1
django-crispy-forms==1.7.0
django-markitup==3.0.0
djangorestframework==3.7.7
docopt==0.6.2
docutils==0.12
flake8==3.5.0
idna==2.6
invoke==0.12.2
isort==4.2.15
Markdown==2.6.11
mccabe==0.6.1
oauthlib==2.0.6
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tox==2.9.1
tox-travis==0.10
urllib3==1.22
virtualenv==15.1.0
